--- android/app/google-services.json
@@ -1,39 +1,0 @@
-{
-  "project_info": {
-    "project_number": "461245967474",
-    "project_id": "pixent",
-    "storage_bucket": "pixent.appspot.com"
-  },
-  "client": [
-    {
-      "client_info": {
-        "mobilesdk_app_id": "1:461245967474:android:ee8cae2cb10cfd95b4a6eb",
-        "android_client_info": {
-          "package_name": "com.example.pixent"
-        }
-      },
-      "oauth_client": [
-        {
-          "client_id": "461245967474-66of4pi58ck4obqi4tuqss3a4aeb4qgn.apps.googleusercontent.com",
-          "client_type": 3
-        }
-      ],
-      "api_key": [
-        {
-          "current_key": "AIzaSyBzdnY4Q0urWLFUwJG0KVNw1HRJlBlUwNg"
-        }
-      ],
-      "services": {
-        "appinvite_service": {
-          "other_platform_oauth_client": [
-            {
-              "client_id": "461245967474-66of4pi58ck4obqi4tuqss3a4aeb4qgn.apps.googleusercontent.com",
-              "client_type": 3
-            }
-          ]
-        }
-      }
-    }
-  ],
-  "configuration_version": "1"
-}
