--- lib/imports/imports.dart
@@ -1,29 +1,0 @@
-export 'package:flutter/material.dart';
-// ignore: depend_on_referenced_packages
-export 'package:flutter_svg/flutter_svg.dart';
-export 'package:get/get.dart';
-export 'package:get/get_core/src/get_main.dart';
-// ignore: depend_on_referenced_packages
-export 'package:google_fonts/google_fonts.dart';
-export 'package:pixent/screen/login/sign_in.dart';
-export 'package:pixent/utility/font/logotext.dart';
-export 'package:pixent/utility/widgets/app_button.dart';
-export 'package:pixent/utility/colors/color.dart';
-export 'package:pixent/screen/login/sign_up.dart';
-export 'package:flutter/material.dart';
-export 'package:get/get_navigation/get_navigation.dart';
-export 'package:pixent/screen/login/login_otion2.dart';
-export 'package:firebase_core/firebase_core.dart';
-export 'package:flutter/material.dart';
-export 'package:pixent/utility/colors/color.dart';
-// ignore: depend_on_referenced_packages
-export 'package:flutter_svg/flutter_svg.dart';
-// ignore: depend_on_referenced_packages
-export 'package:google_fonts/google_fonts.dart';
-export 'package:pixent/utility/font/welcome_text.dart';
-export 'package:slide_action/slide_action.dart';
-export 'package:cupertino_icons/cupertino_icons.dart';
-export 'package:animated_text_kit/animated_text_kit.dart';
-export 'package:google_nav_bar/google_nav_bar.dart';
-export 'package:pixent/imports/imports.dart';
-export 'package:pixent/utility/widgets/background_img_widget.dart';
