+++ lib/main.dart
@@ -1,23 +1,18 @@
-import 'imports/imports.dart';
-//App start from here :-
-Future main() async{
-   WidgetsFlutterBinding.ensureInitialized();
-   await Firebase.initializeApp();
+import 'package:flutter/material.dart';
+import 'package:pixent/screen/login/login_option.dart';
+import 'package:pixent/screen/login_otion2.dart';
 
-   runApp(const MyApp());
+void main() {
+  runApp(MyApp());
 }
+
 class MyApp extends StatelessWidget {
   const MyApp({super.key});
 
   @override
   Widget build(BuildContext context) {
-    return  GetMaterialApp(
-      debugShowCheckedModeBanner: false,      
-      initialRoute: "/",
-      getPages: [
-        GetPage(name: "/", page: (() => const Login_option_()))
-      ],
-      
+    return const MaterialApp(
+      home: Login_option_(),
     );
   }
 }
