--- lib/utility/colors/color.dart
@@ -1,16 +1,0 @@
-import 'dart:ui';
-
-class AppColors{
-  static const Color darkYelloColor = Color(0xffFFBE0B);
-  static const Color lightYellowColor =  Color(0xffFFFE03);
-  static const Color orangeColor =  Color(0xffFB5607);
-  static const Color redColor =  Color(0xffF90201);
-  static const Color purppleColor =  Color(0xff7133FF);
-  static const Color darkPinkColor =  Color(0xffDD0580);
-  static const Color lightPinkColor =  Color(0xffFDC0C0);
-  static const Color skinColor =  Color(0xffFCD0A9);
-  static const Color skeyBlueColor =  Color(0xff1C90F3);
-  static const Color greenDarkColor =  Color(0xff82BC68);
-  static const Color greenLightColor =  Color(0xff00EA86);
-
-}
