--- lib/utility/colors/theme.dart
@@ -1,8 +1,0 @@
-import 'package:flutter/material.dart';
-
-ThemeData themesetup(){
-  return ThemeData(
-    primaryColor: const Color.fromARGB(255, 251, 84, 7),
-    primaryColorDark: const Color.fromARGB(255, 131, 56, 236)
-  );
-}
