+++ pubspec.yaml
@@ -37,13 +37,6 @@ dependencies:
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.2
   slide_action: ^0.0.1+3
-  get: ^4.6.5
-  cloud_firestore: ^3.4.8
-  firebase_auth: ^3.6.2
-  firebase_core: ^1.20.0
-  flutter_screenutil: ^5.5.4
-  animated_text_kit: ^4.2.2
-  google_nav_bar: ^5.0.6
 
 dev_dependencies:
   flutter_test:
@@ -103,10 +96,7 @@ flutter:
     - family: FONTH
       fonts:
         - asset: assets/font/FONTH.ttf
-
-    - family: Modak
-      fonts:
-        - asset: assets/font/Modak-Regular.ttf    
+        
   #     fonts:
   #       - asset: fonts/TrajanPro.ttf
   #       - asset: fonts/TrajanPro_Bold.ttf
